release version 1.0.0 for strcture prediction diffusion

+++ analysis/utils.py
@@ -21,8 +21,8 @@ def create_full_prot(
     assert atom37.shape[-1] == 3
     assert atom37.shape[-2] == 37
     n = atom37.shape[0]
-    residue_index = np.arange(n)
-    chain_index = np.zeros(n)
+    residue_index = np.arange(n,dtype=int)
+    chain_index = np.zeros(n,dtype=int)
     if b_factors is None:
         b_factors = np.zeros([n, 37])
     if aatype is None:
